requirements.txt
pydivert
dnspython